refactor(team): load icons from svg but keep color-change
This change loads the team icons from SVG instead of
including the files directly into the HTML. This has
the advantage of faster page load times since the icons
are not in the first visible view and can therefore be
loaded later. Also for later additions where the icons
will be used in other places as well, they're already
loaded and cached.